Extract the filesystem lowering group into a host submodule
Third cranelift_backend.rs split under #1254, and the first of the
direct-native i64 lowering by capability group. A design pass over the
~18K-line lowering closure (partition analysis across value-shape,
host-vs-core, call-graph-community, and helper lenses) identified the
filesystem family as the cleanest near-leaf: zero callbacks into the
recursive expr/stmt lowering hub.

Move the 26 fs functions (fs_read/fs_write intrinsic lowering, path-guard
and audit helpers, and the compile-time spike_fs_* scope resolvers) plus
the fs-only I64FsReadPath struct into
stage1/crates/axiomc/src/cranelift_backend/host_fs.rs. Two fs-named
functions that sit in the recursive expr SCC
(lower_i64_fs_read_option_match_value_expr, lower_i64_fs_read_some_arm_expr)
stay in the parent. cranelift_backend.rs drops 23,700 -> 22,728 lines and
the top-file share falls below 0.77.

Pure code motion. Shared IR types and the recursive hub stay in the
parent, visible through `use super::*`; the shared I64StaticBindings
struct is raised to pub(crate) for the moved signatures. Call sites are
unchanged via `pub(crate) use host_fs::*`, which also keeps the
already-extracted evaluator.rs sibling's spike_fs_* references resolving.

Refs #1254.
